fix(sync): finalize hierarchical schema migration and theme stabilization
- Migrated all Firestore read/write paths to users/{uid}/drive_sessions subcollections.
- Overhauled MonitorFragment stopMonitoring to construct hierarchical payloads and update global daily analytics via atomic transactions.
- Resolved 'Stale Data' and 'Sync Failure' by implementing persistent real-time SnapshotListeners in Home and Settings tabs.
- Eradicated hardcoded dark hex codes in XML layouts (Edit Profile, Main, Signup) in favor of semantic Material 3 attributes for perfect Light/Dark mode rendering.
- Fixed fatal compilation errors (Unresolved references for AppCompatDelegate and Intent).
- Standardized SystemAnalytics to use daily document IDs (daily_stats_yyyy_MM_dd) with FieldValue.increment logic.

+++ app/src/main/java/com/malik/aegisdrive/model/UserProfile.kt
@@ -1,6 +1,7 @@
 package com.malik.aegisdrive.model
 
 import com.google.firebase.Timestamp
+import com.google.firebase.firestore.Exclude
 import com.google.firebase.firestore.IgnoreExtraProperties
 
 @IgnoreExtraProperties
@@ -16,8 +17,8 @@ data class UserProfile(
 
 @IgnoreExtraProperties
 data class LifetimeStats(
-    val totalDrives: Int = 0,
-    val totalDriveTimeSeconds: Long = 0L,
+    val totalDrives: Long = 0,
+    val totalDriveTimeSeconds: Long = 0,
     val averageSafetyScore: Double = 100.0
 )
